refactor(ui): isolate update schedule projection
